fĩ clonflix

+++ controller/chatController.js
@@ -50,17 +50,24 @@ module.exports = (io) => {
 
             io.to(conversation._id).emit('onmessage', message);
         });
-        //Delete mess event
-        socket.on('deleteMessgae', async data => {
-                await Message.findOneAndDelete({ _id: data.message_id }, (err, docs) => {
-                    if (err) {
-                        socket.emit('deleteMessageError', err)
-                    } else {
-                        socket.emit('deleteMessageSuccess', docs)
-                    }
-                });
-            })
-            //Error handling event
+        socket.on('deleteMessgae',async data=>{
+          let conversation = await Conversation.findOne({
+            "UserList.id": data.userid,
+            "UserList.id": data.partnerid,
+        });
+          if(conversation){
+            await Message.findOneAndDelete({_id:data.message_id},(err,docs)=>{
+              if(err){
+                  io.in(conversation._id).emit('deleteMessageError',err)
+              } else {
+                  io.in(conversation._id).emit('deleteMessageSuccess',docs)
+              }
+          });
+          }
+
+
+        //Error handling event
+
         socket.on('reconnect', (attemptNumber) => {
             console.log(attemptNumber);
         });
